minor corrections to the README.md so that it's instructions agree with the changes made.
Removed the fgmk nameless script in the root folder of the project and changed to a `run.py`. This is what you will run to run the project without installing.
The setup.py now tries to identify what operating system you are using. If it's Ubuntu or Debian, it will add additional entries for icons and mime.